only keep deap squad opt
