+++ airsenal/scripts/squad_builder.py
@@ -4,8 +4,7 @@ import argparse
 import sys
 from typing import List
 
-from airsenal.framework.optimization_deap import make_new_squad_deap
-from airsenal.framework.optimization_squad import Squad, make_new_squad
+from airsenal.framework.optimization_squad import make_new_squad
 from airsenal.framework.optimization_utils import (
     DEFAULT_SUB_WEIGHTS,
     check_tag_valid,
@@ -14,6 +13,7 @@ from airsenal.framework.optimization_utils import (
     get_discounted_squad_score,
 )
 from airsenal.framework.season import CURRENT_SEASON
+from airsenal.framework.squad import Squad
 from airsenal.framework.utils import (
     NEXT_GAMEWEEK,
     fetcher,
@@ -30,12 +30,10 @@ def fill_initial_squad(
     season: str,
     fpl_team_id: int,
     budget: int = 1000,
-    algorithm: str = "genetic",
     remove_zero: bool = True,
     sub_weights: dict = DEFAULT_SUB_WEIGHTS,
     num_generations: int = 100,
     population_size: int = 100,
-    num_iterations: int = 10,
     crossover_prob: float = 0.7,
     mutation_prob: float = 0.3,
     crossover_indpb: float = 0.5,
@@ -44,56 +42,35 @@ def fill_initial_squad(
     verbose: bool = True,
     is_replay: bool = False,  # for replaying seasons
 ) -> Squad:
-    gw_start = gw_range[0]
+    best_squad = make_new_squad(
+        gw_range,
+        tag,
+        budget=budget,
+        season=season,
+        remove_zero=remove_zero,
+        sub_weights=sub_weights,
+        population_size=population_size,
+        generations=num_generations,
+        crossover_prob=crossover_prob,
+        mutation_prob=mutation_prob,
+        crossover_indpb=crossover_indpb,
+        mutation_indpb=mutation_indpb,
+        tournament_size=tournament_size,
+        verbose=verbose,
+    )
 
-    if algorithm == "deap":
-        best_squad = make_new_squad_deap(
-            gw_range,
-            tag,
-            budget=budget,
-            season=season,
-            remove_zero=remove_zero,
-            sub_weights=sub_weights,
-            population_size=population_size,
-            generations=num_generations,
-            crossover_prob=crossover_prob,
-            mutation_prob=mutation_prob,
-            crossover_indpb=crossover_indpb,
-            mutation_indpb=mutation_indpb,
-            tournament_size=tournament_size,
-            verbose=verbose,
-        )
-
-    if algorithm != "deap":
-        if algorithm == "genetic":
-            print(
-                "genetic algorithm not available. Defaulting to algorithm=normal instead"
-            )
-            algorithm = "normal"
-
-        best_squad = make_new_squad(
-            gw_range,
-            tag,
-            budget=budget,
-            season=season,
-            algorithm=algorithm,
-            remove_zero=remove_zero,
-            sub_weights=sub_weights,
-            population_size=population_size,
-            num_iterations=num_iterations,
-            verbose=verbose,
-        )
     if best_squad is None:
         raise RuntimeError(
             "best_squad is None: make_new_squad failed to generate a valid team or "
             "something went wrong with the squad expected points calculation."
         )
 
+    gw_start = gw_range[0]
     optimised_score = get_discounted_squad_score(
         best_squad,
         gw_range,
         tag,
-        gw_range[0],
+        gw_start,
         sub_weights=sub_weights,
     )
     next_points = best_squad.get_expected_points(gw_start, tag)
@@ -137,76 +114,58 @@ def main():
     parser.add_argument(
         "--num_gameweeks", help="how many gameweeks to consider", type=int, default=3
     )
-    parser.add_argument(
-        "--algorithm",
-        help="Which optimization algorithm to use - 'normal' or 'deap'",
-        type=str,
-        default="normal",
-    )
-    # parameters for "normal" optimization
-    parser.add_argument(
-        "--num_iterations",
-        help="number of iterations (normal algorithm only)",
-        type=int,
-        default=10,
-    )
     # parameters for deap optimization
     parser.add_argument(
         "--num_generations",
-        help="number of generations (deap only)",
+        help="number of generations",
         type=int,
         default=100,
     )
     parser.add_argument(
         "--population_size",
-        help="number of candidate solutions per generation (deap only)",
+        help="number of candidate solutions per generation",
         type=int,
         default=100,
     )
     # parameters for "deap" optimization
     parser.add_argument(
         "--crossover_prob",
-        help="crossover probability (deap only)",
+        help="crossover probability",
         type=float,
         default=0.7,
     )
     parser.add_argument(
         "--mutation_prob",
-        help="mutation probability (deap only)",
+        help="mutation probability",
         type=float,
         default=0.3,
     )
     parser.add_argument(
         "--crossover_indpb",
-        help="independent probability for each attribute to be exchanged in crossover (deap only)",
+        help="independent probability for each attribute to be exchanged in crossover",
         type=float,
         default=0.5,
     )
     parser.add_argument(
         "--mutation_indpb",
-        help="independent probability for each attribute to be mutated (deap only)",
+        help="independent probability for each attribute to be mutated",
         type=float,
         default=0.1,
     )
     parser.add_argument(
         "--tournament_size",
-        help="size of tournament for tournament selection (deap only)",
+        help="size of tournament for tournament selection",
         type=int,
         default=3,
     )
     parser.add_argument(
         "--no_subs",
-        help="Don't include points contribution from substitutes (genetic and deap only)",
+        help="Don't include points contribution from substitutes",
         action="store_true",
     )
     parser.add_argument(
         "--include_zero",
-        help="Include players with zero predicted points (genetic and deap only)",
-        action="store_true",
-    )
-    parser.add_argument(
-        "--verbose",
-        help="Print details on optimsation progress",
+        help="Include players with zero predicted points",
         action="store_true",
     )
     parser.add_argument(
@@ -243,8 +202,6 @@ def main():
             "same input gameweeks and season you specified here.",
         )
         sys.exit(1)
-    algorithm = args.algorithm
-    num_iterations = args.num_iterations
     num_generations = args.num_generations
     population_size = args.population_size
     crossover_prob = args.crossover_prob
@@ -253,7 +210,6 @@ def main():
     mutation_indpb = args.mutation_indpb
     tournament_size = args.tournament_size
     remove_zero = not args.include_zero
-    verbose = args.verbose
     fpl_team_id = args.fpl_team_id or fetcher.FPL_TEAM_ID
     if args.no_subs:
         sub_weights = {"GK": 0, "Outfield": (0, 0, 0)}
@@ -266,17 +222,15 @@ def main():
         season=season,
         fpl_team_id=fpl_team_id,
         budget=budget,
-        algorithm=algorithm,
         remove_zero=remove_zero,
         sub_weights=sub_weights,
         num_generations=num_generations,
         population_size=population_size,
-        num_iterations=num_iterations,
         crossover_prob=crossover_prob,
         mutation_prob=mutation_prob,
         crossover_indpb=crossover_indpb,
         mutation_indpb=mutation_indpb,
         tournament_size=tournament_size,
-        verbose=verbose,
+        verbose=True,
         is_replay=args.is_replay,
     )
